Integra las funcionalidades de login y password
