+++ app/src/main/java/com/estebanbe/exploracolombiaapp/controller/TabAdapter.kt
@@ -1,17 +1,14 @@
-package com.estebanbe.exploracolombiaapp.TabAdapter
+package com.estebanbe.exploracolombiaapp.controller
 
 import android.content.Context
 import android.util.Log
-import androidx.core.content.ContentProviderCompat.requireContext
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
-import com.estebanbe.exploracolombiaapp.R
-import com.estebanbe.exploracolombiaapp.TabFragments.ForYouFragment
-import com.estebanbe.exploracolombiaapp.TabFragments.GastronomyFragment
-import com.estebanbe.exploracolombiaapp.TabFragments.HandMadeFragment
-import com.estebanbe.exploracolombiaapp.TabFragments.HotelFragment
-import com.estebanbe.exploracolombiaapp.TabFragments.MapFragment
+import com.estebanbe.exploracolombiaapp.fragments.TabFragments.ForYouFragment
+import com.estebanbe.exploracolombiaapp.fragments.TabFragments.HandMadeFragment
+import com.estebanbe.exploracolombiaapp.fragments.TabFragments.HotelFragment
+import com.estebanbe.exploracolombiaapp.fragments.TabFragments.MapFragment
 
 internal class TabAdapter(var contexto: Context, fm: FragmentManager, var totalTabs: Int): FragmentPagerAdapter(fm) {
