+++ app/src/main/java/com/estebanbe/exploracolombiaapp/fragments/TabFragments/GastroFragments/LiquidGastroFragment.kt
@@ -1,7 +1,6 @@
-package com.estebanbe.exploracolombiaapp.TabFragments
+package com.estebanbe.exploracolombiaapp.fragments.TabFragments.GastroFragments
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -9,14 +8,13 @@ import android.view.ViewGroup
 import com.estebanbe.exploracolombiaapp.R
 
 
-class HandMadeFragment : Fragment() {
+class LiquidGastroFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        Log.e("Fragmento!","HandMade")
-        return inflater.inflate(R.layout.fragment_hand_made, container, false)
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_liquid_gastro, container, false)
     }
-
 }
