+++ app/src/main/java/com/estebanbe/exploracolombiaapp/fragments/TabFragments/GastroFragments/SolidGastroFragment.kt
@@ -1,7 +1,6 @@
-package com.estebanbe.exploracolombiaapp.TabFragments
+package com.estebanbe.exploracolombiaapp.fragments.TabFragments.GastroFragments
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -9,21 +8,14 @@ import android.view.ViewGroup
 import com.estebanbe.exploracolombiaapp.R
 
 
-class HotelFragment : Fragment() {
-
+class SolidGastroFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        Log.e("Fragmento!","Hotel")
-        return inflater.inflate(R.layout.fragment_hotel, container,false)
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_solid_gastro, container, false)
     }
-
-    override fun onStart() {
-        super.onStart()
-        Log.e("Fragment1", "onStart ejecutado")
-    }
-
 
 }
