+++ app/src/main/java/com/estebanbe/exploracolombiaapp/fragments/TabFragments/GastronomyFragment.kt
@@ -1,6 +1,7 @@
-package com.estebanbe.exploracolombiaapp.TabFragments.GastroFragments
+package com.estebanbe.exploracolombiaapp.fragments.TabFragments
 
 import android.os.Bundle
+import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -8,14 +9,12 @@ import android.view.ViewGroup
 import com.estebanbe.exploracolombiaapp.R
 
 
-class SolidGastroFragment : Fragment() {
-
+class GastronomyFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_solid_gastro, container, false)
+        Log.e("Fragmento!","Gastronomia")
+        return inflater.inflate(R.layout.fragment_gastronomy, container, false)
     }
-
 }
