+++ app/src/main/java/com/estebanbe/exploracolombiaapp/fragments/TabFragments/HotelFragment.kt
@@ -1,6 +1,7 @@
-package com.estebanbe.exploracolombiaapp.TabFragments.GastroFragments
+package com.estebanbe.exploracolombiaapp.fragments.TabFragments
 
 import android.os.Bundle
+import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -8,13 +9,21 @@ import android.view.ViewGroup
 import com.estebanbe.exploracolombiaapp.R
 
 
-class LiquidGastroFragment : Fragment() {
+class HotelFragment : Fragment() {
+
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_liquid_gastro, container, false)
+        Log.e("Fragmento!","Hotel")
+        return inflater.inflate(R.layout.fragment_hotel, container,false)
     }
+
+    override fun onStart() {
+        super.onStart()
+        Log.e("Fragment1", "onStart ejecutado")
+    }
+
+
 }
