+++ app/src/main/java/com/estebanbe/exploracolombiaapp/modelo/Offer.kt
@@ -1,4 +1,4 @@
-package com.estebanbe.exploracolombiaapp.TabFragments.GastroFragments.Entities
+package com.estebanbe.exploracolombiaapp.modelo
 
 data class Offer(var uid : String? =null,
                  var imageUrl : String? = null,
