+++ app/src/main/res/layout/fragment_for_you.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TabFragments.ForYouFragment">
+    tools:context=".fragments.TabFragments.ForYouFragment">
 
     <TextView
     android:layout_width="match_parent"
