+++ app/src/main/res/layout/fragment_hand_made.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TabFragments.HandMadeFragment">
+    tools:context=".fragments.TabFragments.HandMadeFragment">
 
     <TextView
         android:layout_width="match_parent"
