+++ app/src/main/res/layout/fragment_liquid_gastro.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TabFragments.GastroFragments.LiquidGastroFragment">
+    tools:context=".fragments.TabFragments.GastroFragments.LiquidGastroFragment">
 
     <TextView
         android:layout_width="match_parent"
