+++ app/src/main/res/layout/fragment_restaurant.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".TabFragments.GastroFragments.RestaurantFragment"
+    tools:context=".fragments.TabFragments.GastroFragments.RestaurantFragment"
     android:layout_marginTop="160dp">
 
     <androidx.recyclerview.widget.RecyclerView
