+++ app/src/main/res/layout/fragment_solid_gastro.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TabFragments.GastroFragments.SolidGastroFragment">
+    tools:context=".fragments.TabFragments.GastroFragments.SolidGastroFragment">
 
     <TextView
         android:layout_width="match_parent"
